# docs/api/openapi-territoires-decisions.yaml
openapi: 3.0.3
info:
  title: API Projets Collectivités — vue territoriale & décisions
  description: >
    Sous-ensemble de l'API communs consommé par MEC (vue territoriale DDT, lecture seule
    + journal de décisions). Voir INTEGRATION_MEC.md pour la sémantique détaillée.
    Les identifiants d'objets (`traces[].id`) sont stables et contractuels ; la composition
    des groupes et la confiance sont recalculées à chaque run du pipeline.
  version: "1.0.0"
security:
  - bearerAuth: []
paths:
  /territoires/{code}/projets:
    get:
      summary: Projets d'un territoire, regroupés par projet réel (déduplication)
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, pattern: '^\d{5}$|^\d{9}$' }
          description: Code INSEE commune (5 chiffres) ou SIREN EPCI (9 chiffres)
        - name: sources
          in: query
          schema: { type: string }
          description: "Filtre source_origine, CSV (ex. `MEC,Vivier COP`)"
        - name: copMillesime
          in: query
          schema: { type: string, enum: ["2024", "2025"] }
        - name: copStatutVivier
          in: query
          schema: { type: string, enum: [a_remonter, a_travailler, hors_cop_mais_crte, non_remonte] }
          description: Statut vivier COP
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: inclureFinancementsSeuls
          in: query
          schema: { type: boolean, default: false }
          description: Inclure les groupes composés uniquement de financements DGCL/Fonds Vert
        - name: masquerObsoletes
          in: query
          schema: { type: boolean, default: false }
          description: Exclure les groupes dont une trace porte une décision active projet_statut / obsolete
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
                  groupes:
                    type: array
                    items: { $ref: "#/components/schemas/Groupe" }
          description: Groupes de traces du territoire
        "404": { description: Territoire inconnu ou code invalide }
  /plans/{cle}/projets-territoire:
    get:
      summary: Projets du territoire couvert par un PCAET (miroir TeT), rattachement par groupe
      description: >
        Miroir de /territoires/{code}/projets côté TeT : part d'un PCAET, liste les projets des
        communes qu'il couvre, regroupés par projet réel, avec par groupe le rattachement à CE PCAET.
        Écriture symétrique via POST /decisions (rattachement_pcaet). Voir INTEGRATION_TET.md.
      parameters:
        - name: cle
          in: path
          required: true
          schema: { type: string }
          description: SIREN du porteur du PCAET (clé stable, 9 chiffres) ou un plan_id de la référence (opendata/snapshot/live)
        - name: sources
          in: query
          schema: { type: string }
          description: "Filtre source_origine, CSV (ex. `MEC,Vivier COP`)"
        - name: copMillesime
          in: query
          schema: { type: string, enum: ["2024", "2025"] }
        - name: copStatutVivier
          in: query
          schema: { type: string, enum: [a_remonter, a_travailler, hors_cop_mais_crte, non_remonte] }
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: inclureFinancementsSeuls
          in: query
          schema: { type: boolean, default: false }
        - name: masquerObsoletes
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  pcaet:
                    type: object
                    properties:
                      sirenPorteur: { type: string }
                      nom: { type: string, nullable: true }
                      source: { type: string, enum: [snapshot, opendata], nullable: true }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
                  groupes:
                    type: array
                    items: { $ref: "#/components/schemas/GroupeRattache" }
          description: Projets du territoire du PCAET, avec rattachement par groupe
        "404": { description: Clé ne résolvant aucun PCAET, ou référence PCAET non matérialisée }
  /projets/mec/{externalId}/plans-territoire:
    get:
      summary: PCAET couvrant les communes d'un projet MEC
      parameters:
        - name: externalId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  pcaet:
                    type: array
                    items:
                      type: object
                      properties:
                        nom: { type: string, nullable: true }
                        sirenPorteur: { type: string }
                        presentDansTet: { type: boolean }
                        tetExternalId: { type: string, nullable: true }
                        source: { type: string, enum: [snapshot, opendata], description: "Canal live exclu en V1 (issue #497)" }
                        rattachement:
                          type: string
                          enum: [confirme, infirme, aucun]
                          description: >
                            Décision active rattachement_pcaet la plus récente entre ce projet et ce PCAET
                            (SIREN porteur) ; 'aucun' par défaut.
                  fichesActionSuggerees:
                    type: array
                    items: { type: object }
                    description: Vide en V1
          description: PCAET du territoire du projet
        "404": { description: externalId MEC inconnu }
  /projets/mec/{externalId}/qualification:
    get:
      summary: Qualification pipeline d'un projet MEC (leviers, thématiques, probabilité TE)
      parameters:
        - name: externalId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  externalId: { type: string }
                  projetId: { type: string, description: ID pivot stable }
                  leviersSgpe:
                    type: array
                    items: { type: string }
                  llmThematiques:
                    type: array
                    nullable: true
                    items:
                      type: object
                      properties:
                        label: { type: string }
                        score: { type: number }
                  llmProbabiliteTe: { type: number, nullable: true }
                  llmClassifiedAt: { type: string, format: date-time, nullable: true }
          description: Qualification
        "404": { description: externalId MEC inconnu }
  /decisions:
    post:
      summary: Enregistrer une décision humaine (journal append-only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDecision" }
      responses:
        "201":
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  createdAt: { type: string, format: date-time }
          description: Décision enregistrée
        "400": { description: Corps invalide }
    get:
      summary: Décisions par objet et/ou par type (audit)
      description: Au moins un filtre requis (objetId et/ou type). Cloisonné au service authentifié.
      parameters:
        - name: objetId
          in: query
          required: false
          schema: { type: string }
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [doublon_signale, doublon_confirme, doublon_infirme, rattachement_pcaet, projet_statut, correction_signalee]
      responses:
        "200": { description: "Décisions du service authentifié uniquement (max 100, plus récentes d'abord)" }
        "400": { description: "Aucun filtre (ni objetId ni type) ou type hors enum" }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  schemas:
    Groupe:
      type: object
      description: >
        Traces d'un même projet réel dans les différentes sources. Aucun identifiant de
        groupe n'est exposé (les identifiants de clusters sont instables entre deux runs).
      properties:
        confiance:
          type: string
          nullable: true
          enum: [CERTAIN, PROBABLE, null]
          description: CERTAIN = lien établi · PROBABLE = suggestion · null = singleton
        traces:
          type: array
          items: { $ref: "#/components/schemas/Trace" }
        decisions:
          type: array
          description: >
            Décisions humaines ACTIVES portant sur une trace du groupe, toutes plateformes
            confondues, SANS l'agent auteur (PII + cloisonnement inter-plateformes). Les révocations
            (verdict='annule') et les décisions supersédées sont exclues.
          items: { $ref: "#/components/schemas/Decision" }
    GroupeRattache:
      description: >
        Groupe augmenté de son rattachement au PCAET interrogé (endpoint /plans/{cle}/projets-territoire).
      allOf:
        - $ref: "#/components/schemas/Groupe"
        - type: object
          properties:
            rattachement:
              type: string
              enum: [confirme, infirme, suggere, aucun]
              description: >
                confirme|infirme = décision active rattachement_pcaet la plus récente entre une trace
                du groupe et ce PCAET (SIREN porteur) ; suggere = pas de décision mais une trace est
                marquée opération PCAET (pcaet_operation_inscrite) — indicatif, une décision prime ;
                aucun = ni décision ni signal.
    Decision:
      type: object
      properties:
        type: { type: string, description: Type de décision (vocabulaire fermé) }
        verdict: { type: string, nullable: true }
        plateforme: { type: string, description: Plateforme émettrice (jamais l'agent auteur) }
        createdAt: { type: string, format: date-time }
        objetAId: { type: string }
        objetBId: { type: string, nullable: true }
        commentaire: { type: string, nullable: true }
    Trace:
      type: object
      properties:
        role: { type: string, enum: [projet, financement] }
        source: { type: string, description: "source_origine (MEC, Vivier COP, DGCL DSIL…)" }
        id: { type: string, description: ID objet STABLE — seule référence à persister }
        externalId: { type: string, nullable: true, description: Renseigné pour les traces MEC }
        nom: { type: string, nullable: true }
        statut: { type: string, nullable: true }
        phase: { type: string, nullable: true }
        budgetPrevisionnel: { type: number, nullable: true, description: Montant demandé / coût du dossier }
        montantAttribue: { type: number, nullable: true, description: Subvention attribuée (traces financement DGCL) }
        copMillesime: { type: string, nullable: true }
        copStatutVivier: { type: string, nullable: true }
    CreateDecision:
      type: object
      required: [typeDecision, objetAType, objetAId]
      properties:
        typeDecision:
          type: string
          enum: [doublon_signale, doublon_confirme, doublon_infirme, rattachement_pcaet, projet_statut, correction_signalee]
          description: >
            Vocabulaire FERMÉ (400 sinon). Contraintes croisées par type sur objetB/verdict/payload
            (voir GUIDE_DECISIONS.md). Ex. : verdict interdit pour les doublons ; requis pour
            rattachement_pcaet (confirme|infirme) et projet_statut (valide|obsolete|termine) ;
            payload { champ, valeurProposee, source? } requis pour correction_signalee.
        objetAType: { type: string, enum: [projet, fiche_action, plan, financement] }
        objetAId: { type: string, description: ID objet stable (JAMAIS un identifiant de groupe) }
        objetBType:
          type: string
          enum: [projet, fiche_action, plan, financement, pcaet]
          nullable: true
          description: "'pcaet' réservé à rattachement_pcaet (objetBId = SIREN porteur, 9 chiffres)"
        objetBId: { type: string, nullable: true }
        verdict:
          type: string
          nullable: true
          description: >
            Valeurs contraintes par type (voir typeDecision). Cas spécial : 'annule' est valide pour
            TOUS les types SI supersedes est fourni (400 sinon) — révocation qui retire la cible sans
            rien affirmer ; les lignes verdict='annule' sont exclues de tous les effets de lecture.
        auteur: { type: string, nullable: true }
        commentaire: { type: string, nullable: true, description: "Exposé cross-plateforme dans decisions[] — pas de PII" }
        payload: { type: object, nullable: true, description: "Max ~10 ko sérialisé ; forme imposée pour correction_signalee" }
        supersedes:
          type: string
          format: uuid
          nullable: true
          description: "Décision révoquée. Doit être de la MÊME plateforme et du MÊME typeDecision (400 sinon)."
      description: >
        plateforme_source n'est PAS dans le corps — elle est dérivée du service authentifié.